Handle /user page while lacking session Fixes #69
