+++ frontend/src/components/LoginFormPage/index.js
@@ -12,9 +12,9 @@ export default function LoginFormPage() {
 	const [password, setPassword] = useState("");
 	const [errors, setErrors] = useState([]);
 
-	if (sessionUser) return (
-		<Redirect to="/user" />
-	)
+	if (sessionUser) {
+		return (<Redirect to="/user" />)
+	}
 
 	function handleSubmission(e) {
 		e.preventDefault();
